requirements.txt
aiohttp==3.6.3
async-timeout==3.0.1
attrs==20.2.0
beautifulsoup4==4.9.3
bs4==0.0.1
chardet==3.0.4
discord.py==1.5.1
idna==2.10
multidict==4.7.6
soupsieve==2.0.1
yarl==1.5.1